+++ src/consts/HTTPStatusCodes.ts
@@ -1,0 +1,5 @@
+export enum HTTPStatusCodes {
+  BAD_REQUEST = 400,
+  NOT_FOUND = 404,
+  INTERNAL_SERVER_ERROR = 500,
+}
