+++ src/consts/globals.ts
@@ -1,0 +1,2 @@
+// Whether we're in development mode or not
+export const IS_DEV_MODE = process.env.NODE_ENV === "development";
