+++ src/services/models/responses/Response.model.ts
@@ -1,4 +1,4 @@
-import type { ID } from "../../CoreTypes";
+import type { ID } from "../../../consts/ids";
 import ModelBase from "../ModelBase";
 
 export type TResponse<TInput, TEvaluation> = {
